--- src/main/java/ToDo.java
@@ -1,10 +1,0 @@
-public class ToDo extends Task {
-    public ToDo(String desc) {
-        super(desc, TaskType.TODO);
-    }
-
-    @Override
-    public String toString() {
-        return getType() + super.toString();
-    }
-}
