+++ src/main/java/connor/Parser.java
@@ -1,3 +1,9 @@
+package connor;
+
+import connor.command.*;
+import connor.task.TaskStatus;
+import connor.task.TaskType;
+
 public class Parser {
     private static final String ERROR_INDEX_NOT_INTEGER = "Error! Index must be a valid integer.";
     private static final String ERROR_INVALID_COMMAND_START = "My apologies, I don't understand what '";
