+++ src/main/java/connor/command/ByeCommand.java
@@ -1,3 +1,7 @@
+package connor.command;
+
+import connor.Connor;
+
 public class ByeCommand extends Command {
 
     public ByeCommand() {}
