+++ src/main/java/connor/command/ClearCommand.java
@@ -1,3 +1,7 @@
+package connor.command;
+
+import connor.task.TaskList;
+
 public class ClearCommand extends Command {
     public ClearCommand() {}
