+++ src/main/java/connor/command/Command.java
@@ -1,3 +1,5 @@
+package connor.command;
+
 public abstract class Command {
 
     public abstract void activate();
