+++ src/main/java/connor/command/CommandType.java
@@ -1,3 +1,5 @@
+package connor.command;
+
 public enum CommandType {
     BYE, LIST, ADD, DELETE, CLEAR, CHANGE_STATUS, UNKNOWN;
 }
