+++ src/main/java/connor/exception/ConnorException.java
@@ -1,3 +1,5 @@
+package connor.exception;
+
 public class ConnorException extends Exception {
     public ConnorException(String message) {
         super(message);
