+++ src/main/java/connor/exception/InvalidTaskFileException.java
@@ -1,3 +1,5 @@
+package connor.exception;
+
 public class InvalidTaskFileException extends ConnorException {
     public InvalidTaskFileException(String message) {
         super(message);
