+++ src/main/java/connor/task/Task.java
@@ -1,3 +1,5 @@
+package connor.task;
+
 public class Task {
     protected String desc;
     protected boolean isDone = false;
