+++ src/main/java/connor/task/TaskStatus.java
@@ -1,3 +1,5 @@
+package connor.task;
+
 public enum TaskStatus {
     MARK, UNMARK
 }
