+++ src/main/java/connor/task/TaskType.java
@@ -1,3 +1,5 @@
+package connor.task;
+
 public enum TaskType {
     TODO, DEADLINE, EVENT;
 }
