+++ src/main/java/connor/task/ToDo.java
@@ -1,0 +1,12 @@
+package connor.task;
+
+public class ToDo extends Task {
+    public ToDo(String desc) {
+        super(desc, TaskType.TODO);
+    }
+
+    @Override
+    public String toString() {
+        return getType() + super.toString();
+    }
+}
